Add funny testing comment

+++ Source/Boot/Kernel.cpp
@@ -50,7 +50,7 @@ void Kernel::Initialize()
 void Kernel::Execute()
 {
   typedef void (*KernelEntry)(UInt32, UInt32);
-  ((KernelEntry)s_KernelEntry)(MULTIBOOT_EAX_MAGIC, (UInt32)&s_Header);
+  ((KernelEntry)s_KernelEntry)(MULTIBOOT_EAX_MAGIC, (UInt32)(UInt64)&s_Header); // TODO: Check if this actually works lol
 
   while (true);
 }
